⚠️ | Important.md
Read this file carefully, pre-cautions, intended-use and warnings need to be taken in consideration before implementing this datapack. I can not garantuee it is gonna be problem-free. ⚠️ | Any damages caused by my datapack I will not be responsible for, unless I did not mention this in this file. Otherwise, if you did something that was advised against, I will not be taking responsibility.